Support unauthenticated users in hunt invitation flow.
Unauthenticated users who open a /join URL will see a welcome screen
asking whether they are a new or existing user. New users are switched
to the enrollment screen; existing users see the login screen as before.

The enrollment screen is updated to support brand new account creation,
rather than just setting a password for an already-created account. If
an invitation code is present, the user can enter an email, or elect to
sign in with a Google account which can be used for future sign ins. As
before, they can set initial profile fields as well. They will be added
to the hunt once they are signed in (by virtue of being redirected to
the original /join link from location state, but now as an authenticated
user).

If a new user goes through the Google sign-in flow, the email and
password fields are subsequently hidden and the credentials stashed
until the user enters the remaining profile fields and completes
enrollment.