Exclude Drive Activity from the server's Google account.
It appears that the drive API counts various actions taken by the server
account in the process of creating a new document for a puzzle as edits
(in that the edit field is populated). This is confusing as it looks
like there is drive activity in every new puzzle when it is first
created, even if nobody has opened it or edited the document.

We can filter out events associated with the ID of the server account to
prevent this confusion. Given that this account is recommended to not be
one used for actual hunting, this seems like a reasonable approach.

We were not previously storing the ID when logging into a server
account. We store it for future sign-ins, but for simplicity, do not
bother with backfill. This means that until the server admin signs in
again, they will continue to see activity under the server admin.

Also note that regardless of which account you're logged into JR with,
Sheets activity will be associated with the default Google account in
your browser session.