Add new APIs for creating puzzles.
- Propagate user ID to API methods in authenticator
- Add API to list hunts
- Add API to list tags for a hunt (for autocompletion)
- Add API to create a puzzle

Most of the complexity is in the latter. The bulk of the implementation
is shared with the existing flow for creating puzzles, but there is some
duplicate input type checking in both the Meteor method (to take
advantage of the validate scaffolding) as well as the shared logic (so
it's there for API calls).

See #237
See #1948