Downgrade MediaSoup to 3.11.2.
This is an attempt at working around #2169. The bug indicates loops of
tens of thousands of ChannelMessageRegistrator::RegisterHandler failures
all for the same ID due to the handler already existing. In 3.11.2, the
first failure would cause the existing handler to be unregistered, which
would presumably make the second attempt succeed. In 3.11.3, it the
existing handler would be left around, which could certainly explain a
loop of repeated errors as the state isn't changing.

It seems likely that there is still another latent bug causing handlers
to be abandoned instead of being unregistered properly, but perhaps
going back to the old behavior is a better compromise until this can be
fixed properly.

See #2169